Move parse number and range extract funcs
